Move app specific components into $lib/ui
