+++ src/lib/ui/util/modal.ts
@@ -1,4 +1,4 @@
-import type { ModalContext } from '../modal';
+import type { ModalContext } from 'packages/ui/modal';
 import ErrorDialog from './ErrorDialog.svelte';
 
 export async function displayErrorDialog(modal: ModalContext, error: unknown) {
